[java_client] Moving files and updated the imports
