+++ java_client/ChitChatDesktop/src/com/moc/chitchat/controller/UserSearchController.java
@@ -1,4 +1,4 @@
-package com.moc.chitchat.controller.authentication;
+package com.moc.chitchat.controller;
 
 import com.mashape.unirest.http.HttpResponse;
 import com.mashape.unirest.http.JsonNode;
